+++ NEWS.md
@@ -1,3 +1,5 @@
+# spanishoddata (development version)
+
 # spanishoddata 0.2.5
 
 ## Bug fixes